+++ src/dune_rules/obj_dir.ml
@@ -22,24 +22,32 @@ module External = struct
   type t =
     { public_dir : Path.t
     ; private_dir : Path.t option
+    ; public_cmi_dir : Path.t option
     }
 
-  let make ~dir ~has_private_modules =
+  let make ~dir ~has_private_modules ~private_lib =
     let private_dir =
       if has_private_modules then
         Some (Path.relative dir ".private")
       else
         None
     in
-    { public_dir = dir; private_dir }
+    let public_cmi_dir =
+      if private_lib then
+        Some (Path.relative dir ".public_cmi")
+      else
+        None
+    in
+    { public_dir = dir; private_dir; public_cmi_dir }
 
-  let public_cmi_dir t = t.public_dir
+  let public_cmi_dir t = Option.value ~default:t.public_dir t.public_cmi_dir
 
-  let to_dyn { public_dir; private_dir } =
+  let to_dyn { public_dir; private_dir; public_cmi_dir } =
     let open Dyn.Encoder in
     record
       [ ("public_dir", Path.to_dyn public_dir)
       ; ("private_dir", option Path.to_dyn private_dir)
+      ; ("public_cmi_dir", option Path.to_dyn public_cmi_dir)
       ]
 
   let cm_dir t (cm_kind : Cm_kind.t) (visibility : Visibility.t) =
@@ -47,25 +55,30 @@ module External = struct
     | Cmi, Private, Some p -> p
     | Cmi, Private, None ->
       Code_error.raise "External.cm_dir" [ ("t", to_dyn t) ]
-    | Cmi, Public, _
-    | (Cmo | Cmx), _, _ ->
-      t.public_dir
+    | Cmi, Public, _ -> public_cmi_dir t
+    | (Cmo | Cmx), _, _ -> t.public_dir
 
-  let encode { public_dir; private_dir } =
+  let encode { public_dir; private_dir; public_cmi_dir } =
     let open Dune_lang.Encoder in
     let extract d =
       Path.descendant ~of_:public_dir d |> Option.value_exn |> Path.to_string
     in
     let private_dir = Option.map ~f:extract private_dir in
-    record_fields [ field_o "private_dir" string private_dir ]
+    let public_cmi_dir = Option.map ~f:extract public_cmi_dir in
+    record_fields
+      [ field_o "private_dir" string private_dir
+      ; field_o "public_cmi_dir" string public_cmi_dir
+      ]
 
   let decode ~dir =
+    let public_dir = dir in
     let open Dune_lang.Decoder in
     fields
-      (let+ private_dir = field_o "private_dir" string in
-       let public_dir = dir in
+      (let+ private_dir = field_o "private_dir" string
+       and+ public_cmi_dir = field_o "public_cmi_dir" string in
        let private_dir = Option.map ~f:(Path.relative dir) private_dir in
-       { public_dir; private_dir })
+       let public_cmi_dir = Option.map ~f:(Path.relative dir) public_cmi_dir in
+       { public_dir; private_dir; public_cmi_dir })
 
   let byte_dir t = t.public_dir
 
@@ -79,8 +92,8 @@ module External = struct
 
   let all_obj_dirs t ~mode:_ = [ t.public_dir ]
 
-  let all_cmis { public_dir; private_dir } =
-    List.filter_opt [ Some public_dir; private_dir ]
+  let all_cmis { public_dir; private_dir; public_cmi_dir } =
+    List.filter_opt [ Some public_dir; private_dir; public_cmi_dir ]
 
   let cm_public_dir t (cm_kind : Cm_kind.t) =
     match cm_kind with
@@ -96,9 +109,11 @@ module Local = struct
     ; native_dir : Path.Build.t
     ; byte_dir : Path.Build.t
     ; public_cmi_dir : Path.Build.t option
+    ; private_lib : bool
     }
 
-  let to_dyn { dir; obj_dir; native_dir; byte_dir; public_cmi_dir } =
+  let to_dyn { dir; obj_dir; native_dir; byte_dir; public_cmi_dir; private_lib }
+      =
     let open Dyn.Encoder in
     record
       [ ("dir", Path.Build.to_dyn dir)
@@ -106,10 +121,11 @@ module Local = struct
       ; ("native_dir", Path.Build.to_dyn native_dir)
       ; ("byte_dir", Path.Build.to_dyn byte_dir)
       ; ("public_cmi_dir", option Path.Build.to_dyn public_cmi_dir)
+      ; ("private_lib", bool private_lib)
       ]
 
-  let make ~dir ~obj_dir ~native_dir ~byte_dir ~public_cmi_dir =
-    { dir; obj_dir; native_dir; byte_dir; public_cmi_dir }
+  let make ~dir ~obj_dir ~native_dir ~byte_dir ~public_cmi_dir ~private_lib =
+    { dir; obj_dir; native_dir; byte_dir; public_cmi_dir; private_lib }
 
   let need_dedicated_public_dir t = Option.is_some t.public_cmi_dir
 
@@ -134,7 +150,7 @@ module Local = struct
     in
     Path.Build.Set.of_list dirs |> Path.Build.Set.to_list
 
-  let make_lib ~dir ~has_private_modules lib_name =
+  let make_lib ~dir ~has_private_modules ~private_lib lib_name =
     let obj_dir = Paths.library_object_directory ~dir lib_name in
     let public_cmi_dir =
       Option.some_if has_private_modules (Paths.library_public_cmi_dir ~obj_dir)
@@ -142,14 +158,14 @@ module Local = struct
     make ~dir ~obj_dir
       ~native_dir:(Paths.library_native_dir ~obj_dir)
       ~byte_dir:(Paths.library_byte_dir ~obj_dir)
-      ~public_cmi_dir
+      ~public_cmi_dir ~private_lib
 
   let make_exe ~dir ~name =
     let obj_dir = Paths.executable_object_directory ~dir name in
     make ~dir ~obj_dir
       ~native_dir:(Paths.library_native_dir ~obj_dir)
       ~byte_dir:(Paths.library_byte_dir ~obj_dir)
-      ~public_cmi_dir:None
+      ~public_cmi_dir:None ~private_lib:false
 
   let cm_dir t cm_kind _ =
     match cm_kind with
@@ -190,11 +206,11 @@ let decode ~dir =
   let+ external_ = External.decode ~dir in
   External external_
 
-let make_lib ~dir ~has_private_modules lib_name =
-  Local (Local.make_lib ~dir ~has_private_modules lib_name)
+let make_lib ~dir ~has_private_modules ~private_lib lib_name =
+  Local (Local.make_lib ~dir ~has_private_modules ~private_lib lib_name)
 
 let make_external_no_private ~dir =
-  External (External.make ~dir ~has_private_modules:false)
+  External (External.make ~dir ~has_private_modules:false ~private_lib:false)
 
 let get_path :
     type a. a t -> l:(Local.t -> Path.Build.t) -> e:(External.t -> Path.t) -> a
@@ -226,7 +242,8 @@ let convert_to_external (t : Path.Build.t t) ~dir =
   match t with
   | Local e ->
     let has_private_modules = Local.need_dedicated_public_dir e in
-    External (External.make ~dir ~has_private_modules)
+    External
+      (External.make ~dir ~has_private_modules ~private_lib:e.private_lib)
   | _ -> assert false
 
 let all_cmis (type path) (t : path t) : path list =